Changed styles
