+++ frontend/sozisel-app/src/components/MainNavbar/MainNavbar.scss
@@ -28,6 +28,7 @@
         margin-top: $spacing-01;
         margin-bottom: 0;
         list-style-type: none;
+        box-shadow: 0px 0px 40px rgba(0, 0, 0, 0.1);
         li {
             @include disableLinkDecoration();
             margin-left: $spacing-07;
