+++ frontend/sozisel-app/src/components/TemplatesList/SearchBar/SearchBar.scss
@@ -1,15 +1,11 @@
 @import "../../../styles/styles";
 
 .searchBar{
-    padding: 10px 50px 10px 50px;
+    padding: $spacing-04 $spacing-09;
 
     .gridItem{
         display: flex;
         align-items: center;
-    }
-    
-    .searchBarButton{
-        padding: 15px 10px 15px 10px;
     }
 }
