+++ frontend/sozisel-app/src/components/TemplatesList/SearchBar/SearchBar.tsx
@@ -1,6 +1,5 @@
 import "./SearchBar.scss";
 import { BaseSyntheticEvent, ReactElement } from "react";
-import FilledInput from "@material-ui/core/FilledInput";
 import InputAdornment from "@material-ui/core/InputAdornment";
 import FormControl from "@material-ui/core/FormControl";
 import { useState } from "react";
@@ -10,9 +9,9 @@ import ClearIcon from "@material-ui/icons/Clear";
 import MenuItem from "@material-ui/core/MenuItem";
 import Select from "@material-ui/core/Select";
 import Button from "@material-ui/core/Button";
-import AddIcon from "@material-ui/icons/Add";
 import { useTranslation } from "react-i18next";
 import { useHistory } from "react-router-dom";
+import TextField from "@material-ui/core/TextField";
 
 export interface SearchBarProps {
   onSearch: ({
@@ -55,19 +54,20 @@ export default function SearchBar({ onSearch }: SearchBarProps): ReactElement {
     <>
       <Grid container spacing={3} justify="center" className="searchBar">
         <Grid item sm={12} md={4} className="gridItem">
-          <FormControl fullWidth variant="filled">
-            <FilledInput
-              id="searchTextInput"
-              value={name}
-              disableUnderline
-              inputProps={{ style: { padding: "20px 10px 20px 10px" } }}
-              onChange={onSearchTextChange}
-              startAdornment={
+          <TextField
+            fullWidth
+            id="searchTextInput"
+            variant="outlined"
+            size="small"
+            value={name}
+            onChange={onSearchTextChange}
+            InputProps={{
+              startAdornment: (
                 <InputAdornment position="end">
                   <SearchIcon />
                 </InputAdornment>
-              }
-              endAdornment={
+              ),
+              endAdornment: (
                 <InputAdornment position="end">
                   {name != "" && (
                     <ClearIcon
@@ -77,30 +77,28 @@ export default function SearchBar({ onSearch }: SearchBarProps): ReactElement {
                     />
                   )}
                 </InputAdornment>
-              }
-            />
-          </FormControl>
+              ),
+            }}
+          />
         </Grid>
         <Grid item sm={12} md={4} className="gridItem">
-          <FormControl fullWidth variant="filled">
-            <Select
-              disableUnderline
-              autoWidth
-              SelectDisplayProps={{
-                style: { padding: "20px 50px 20px 10px" },
-              }}
-              id="accessSelect"
-              value={includePublic}
-              onChange={onSearchPublicChange}
-            >
-              <MenuItem value={0}>
-                {t("components.TemplatesList.private")}
-              </MenuItem>
-              <MenuItem value={1}>
-                {t("components.TemplatesList.public")}
-              </MenuItem>
-            </Select>
-          </FormControl>
+          <Select
+            fullWidth
+            variant="outlined"
+            SelectDisplayProps={{
+              style: { padding: "11px" },
+            }}
+            id="accessSelect"
+            value={includePublic}
+            onChange={onSearchPublicChange}
+          >
+            <MenuItem value={0}>
+              {t("components.TemplatesList.private")}
+            </MenuItem>
+            <MenuItem value={1}>
+              {t("components.TemplatesList.public")}
+            </MenuItem>
+          </Select>
         </Grid>
         <Grid item sm={6} md={2} className="gridItem">
           <Button
@@ -118,7 +116,6 @@ export default function SearchBar({ onSearch }: SearchBarProps): ReactElement {
             variant="contained"
             color="primary"
             fullWidth
-            startIcon={<AddIcon />}
             className="searchBarButton"
             onClick={() => history.push("/templates/create")}
           >
