+++ frontend/sozisel-app/src/styles/color.scss
@@ -1,2 +1,2 @@
 $main-color: #ff9800;
-$bg-01: #f0f0f0;
+$bg-01: #ffffff;
