+++ frontend/sozisel-app/src/styles/spacing.scss
@@ -6,4 +6,5 @@ $spacing-05: 20px;
 $spacing-06: 26px;
 $spacing-07: 32px;
 $spacing-08: 40px;
-$spacing-09: 60px;
+$spacing-09: 50px;
+$spacing-10: 60px;
